# Improved performance of gizmo curve rendering in 2.9