--- a/com.unity.cinemachine/Runtime/Behaviours/CinemachinePath.cs
+++ b/com.unity.cinemachine/Runtime/Behaviours/CinemachinePath.cs
@@ -98,10 +98,17 @@ namespace Cinemachine
             return pos;
         }
 
+#if CINEMACHINE_USE_BATCH_POINT_LINE_APIS
         /// <summary>Get a curvespace position of a point along the path</summary>
         /// <param name="pos">Postion along the path.  Need not be normalized.</param>
         /// <returns>Curve-space position of the point along at path at pos</returns>
         public override Vector3 EvaluateCurvePosition(float pos)
+#else
+        /// <summary>Get a worldspace position of a point along the path</summary>
+        /// <param name="pos">Postion along the path.  Need not be normalized.</param>
+        /// <returns>World-space position of the point along at path at pos</returns>
+        public override Vector3 EvaluatePosition(float pos)
+#endif
         {
             Vector3 result = new Vector3();
             if (m_Waypoints.Length == 0)
@@ -122,9 +129,14 @@ namespace Cinemachine
                         wpB.position - wpB.tangent, wpB.position);
                 }
             }
+#if CINEMACHINE_USE_BATCH_POINT_LINE_APIS
             return result;
+#else
+            return transform.TransformPoint(result);
+#endif
         }
 
+#if CINEMACHINE_USE_BATCH_POINT_LINE_APIS
         /// <summary>Get a worldspace position of a point along the path</summary>
         /// <param name="pos">Postion along the path.  Need not be normalized.</param>
         /// <returns>World-space position of the point along at path at pos</returns>
@@ -132,6 +144,7 @@ namespace Cinemachine
         {
             return transform.TransformPoint(EvaluateCurvePosition(pos));
         }
+#endif
 
         /// <summary>Get the tangent of the curve at a point along the path.</summary>
         /// <param name="pos">Postion along the path.  Need not be normalized.</param>