… com.unity.cinemachine/Runtime/Core/SplineHelpers.cs