### Purpose of this PR

While profiling slow editor scene view performance in the _Volvo Test Track_ project it was found that a large proportion of the time was being spent rendering the Cinemachine (2.9) curves as part of gizmo rendering.  To improve the performance of such operations some new batch APIs have been added to Unity allowing multiple points to be transformed or multiple lines drawn by a single API call:

**NOTE: This PR must not land before the Unity PR "Added batch versions of Transform.TransformXXX and Gizmo.Drawline APIs [2022.2]" (https://github.cds.internal.unity3d.com/unity/unity/pull/18472) does - it is open purely for review at this time.**

Batch versions of Transform.TransformPoint

- Transform.TransformPoints(Span<Vector3> positions)
- Transform.TransformPoints(ReadOnlySpan<Vector3> positions, Span<Vector3> transformedPositions)- 

Batch version of Transform.InverseTransformPoint

- Transform.InverseTransformPoints(Span<Vector3> positions)
- Transform.InverseTransformPoints(ReadOnlySpan<Vector3> positions, Span<Vector3> transformedPositions)- 

Batch version of Transform.TransformVector

- Transform.TransformVectors(Span<Vector3> vectors)
- Transform.TransformVectors(ReadOnlySpan<Vector3> vectors, Span<Vector3> transformedVectors)- 

Batch version of Transform.InverseTransformVector

- Transform.InverseTransformVectors(Span<Vector3> vectors)
- Transform.InverseTransformVectors(ReadOnlySpan<Vector3> vectors, Span<Vector3> transformedVectors)- 

Batch version of Transform.TransformDirection

- Transform.TransformDirections(Span<Vector3> directions)
- Transform.TransformDirections(ReadOnlySpan<Vector3> directions, Span<Vector3> Directions)- 

Batch version of Transform.InverseTransformDirection

- Transform.InverseTransformDirections(Span<Vector3> directions)
- Transform.InverseTransformDirections(ReadOnlySpan<Vector3> directions, Span<Vector3> Directions)- 

Trunk PR (landed): https://github.cds.internal.unity3d.com/unity/unity/pull/15733
2022.2 PR (under review): https://github.cds.internal.unity3d.com/unity/unity/pull/18472

This PR adds support for these APIs to Cinemachine 2.9 more than doubling the performance of it's gizmo curve rendering.  

A version define CINEMACHINE_USE_BATCH_POINT_LINE_APIS has been added to the editor and runtime asmdefs allowing the new code to only apply to the current 2022.2.0b16 version and later, the original gizmo curve rendering code remains in place for use when the package is used with earlier Unity versions.

### Testing status

Functional tests for the API additions were part of the original PRs.  Manual testing of the Cinemachine changes in this PR was done using one of the sample scenes included with the CM package plus the _Volvo Test Track_ project to make sure that the curves looked the same before and after then change.

![image](https://user-images.githubusercontent.com/35338585/203091946-0240b972-f8f8-4650-846f-49f561c60779.png)
![image](https://user-images.githubusercontent.com/35338585/203092152-bf8f5a7c-af66-4407-822b-08f22e9856ca.png)
![image](https://user-images.githubusercontent.com/35338585/203092651-037402c1-4fcb-4a0f-9b41-cc868a8ed0fa.png)
![image](https://user-images.githubusercontent.com/35338585/203092669-65ad734a-2622-477f-b795-6702a053bc61.png)
![image](https://user-images.githubusercontent.com/35338585/203092678-ddf65b53-0ad3-40c6-b6ee-382b25093969.png)
![image](https://user-images.githubusercontent.com/35338585/203092692-2b0dc496-a8d1-47ff-bebe-d11a584b29d3.png)

A non-automated performance test was used to verify the improvement in scene rendering performance, taking the median of ten redraws in the _Volvo Test Track_ project the scene view render time reduced from 181ms to 89ms.  

Looking more in depth using the Superluminal profiler it was shown that `CinemachineSmoothPathEditor.DrawGizmos()` in this scene reduced from 135ms to 28ms (so over 4.8x the speed)

Backwards compatibility testing of the `CINEMACHINE_USE_BATCH_POINT_LINE_APIS `version define was done by manually changing the asmdef so it only applied to a future version then re-testing the Cinemachine sample scene and Volvo Test Track project to ensure curve rendering was unaffected (apart from performance which was verified to be slower once again)

- [*] Added an automated test
- [*] Passed all automated tests
- [*] Manually tested 

### Documentation status

Is purely a performance improvement, there is no functional change so it wasn't thought that documentation changes were necessary

### Technical risk

Functional tests for the API additions were part of the original PRs.  Manual functional testing of curve rendering was also performed with the Volvo Test Track project as described above to minimise technical risk

There is limited halo risk as only the rendering of curves is affected, not their function or any user interface code

### Comments to reviewers

### Package version

[Justification for updating either the patch, minor, or major version according to the [semantic versioning](https://semver.org/spec/v2.0.0.html) rules]

- [ ] Updated package version