@John-Whigham 2022-11-22
> 

Drawing a mesh may well be faster if it can be cached (I suspect building it each time would not be faster than the code with these new APIs, may or may not be faster than the original method), certainly worth testing. 

The points for the curves using the Gizmo Line APIs could also be cached of course to significantly improve performance with this method (at the expense of memory) but I didn't feel confident enough in my understandings of CM to implement that - particularly knowing when it had to be regenerated