@John-Whigham 2022-11-22
Note this PR won't be proceeding, @glabute is going to tackle the performance issue by a different route generating and caching local space meshes for the curve rendering which should give good performance without requiring the new APIs.