@glabute 2022-11-29
Implemented by #674